[FIX] ProductDTO, also deleted SolutionDTO since nothing used it and it was useless (it was an idea from the start that seems to be not used)
