+++ Eventy/src/main/java/org/example/eventy/solutions/dtos/ProductDTO.java
@@ -8,28 +8,43 @@ import org.example.eventy.solutions.models.Product;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ProductDTO extends SolutionDTO {
+public class ProductDTO {
     private Long id;
+    private String name;
+    private String description;
+    private double price;
+    private double discount;
+    private List<EventTypeDTO> relatedEventTypes;
+    private List<String> images;
+    private boolean isVisible;
+    private boolean isAvailable;
 
     public ProductDTO() {
         super();
     }
 
-    public ProductDTO(Long id, String name, String description, double price, double discount, List<EventTypeDTO> relatedEventTypes, List<String> images, boolean visibility, boolean availability) {
-        super(name, description, price, discount, relatedEventTypes, images, visibility, availability);
+    public ProductDTO(Long id, String name, String description, double price, double discount, List<EventTypeDTO> relatedEventTypes, List<String> images, boolean isVisible, boolean isAvailable) {
         this.id = id;
+        this.name = name;
+        this.description = description;
+        this.price = price;
+        this.discount = discount;
+        this.relatedEventTypes = relatedEventTypes;
+        this.images = images;
+        this.isVisible = isVisible;
+        this.isAvailable = isAvailable;
     }
 
     public ProductDTO(Product product) {
         this.id = product.getId();
-        super.setName(product.getName());
-        super.setDescription(product.getDescription());
-        super.setPrice(product.getPrice());
-        super.setDiscount(product.getDiscount());
-        super.setRelatedEventTypes(product.getEventTypes().stream().map(EventTypeDTO::new).collect(Collectors.toList()));
-        super.setImages(product.getImageUrls().stream().map(PicturePath::getPath).map(PictureService::getImage).collect(Collectors.toList()));
-        super.setAvailability(product.isAvailable());
-        super.setVisibility(product.isVisible());
+        this.setName(product.getName());
+        this.setDescription(product.getDescription());
+        this.setPrice(product.getPrice());
+        this.setDiscount(product.getDiscount());
+        this.setRelatedEventTypes(product.getEventTypes().stream().map(EventTypeDTO::new).collect(Collectors.toList()));
+        this.setImages(product.getImageUrls().stream().map(PicturePath::getPath).map(PictureService::getImage).collect(Collectors.toList()));
+        this.setAvailable(product.isAvailable());
+        this.setVisible(product.isVisible());
     }
 
     public Long getId() {
@@ -38,5 +53,69 @@ public class ProductDTO extends SolutionDTO {
 
     public void setId(Long id) {
         this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public double getDiscount() {
+        return discount;
+    }
+
+    public void setDiscount(double discount) {
+        this.discount = discount;
+    }
+
+    public List<EventTypeDTO> getRelatedEventTypes() {
+        return relatedEventTypes;
+    }
+
+    public void setRelatedEventTypes(List<EventTypeDTO> relatedEventTypes) {
+        this.relatedEventTypes = relatedEventTypes;
+    }
+
+    public List<String> getImages() {
+        return images;
+    }
+
+    public void setImages(List<String> images) {
+        this.images = images;
+    }
+
+    public boolean isVisible() {
+        return isVisible;
+    }
+
+    public void setVisible(boolean visible) {
+        this.isVisible = visible;
+    }
+
+    public boolean isAvailable() {
+        return isAvailable;
+    }
+
+    public void setAvailable(boolean available) {
+        this.isAvailable = available;
     }
 }
